+++ src/modules/scenarioeditor/ChartEditView.vue
@@ -41,6 +41,9 @@
         class="absolute top-2 left-2 z-10 bg-gray-50 bg-opacity-80"
         :items="breadcrumbItems"
       />
+      <nav class="absolute right-4 top-2 z-10 rounded-full bg-white">
+        <DotsMenu :items="menuItems" />
+      </nav>
 
       <ToggleField class="absolute bottom-2 right-2 z-10" v-model="debug"
         >Debug mode</ToggleField
@@ -55,7 +58,7 @@
         :options="options.$state"
         :specific-options="specificOptions.$state"
         enable-pan-zoom
-        interactive
+        :interactive="isInteractive"
         @unitclick="onUnitClick"
         @levelclick="onLevelClick"
         @branchclick="onBranchClick"
@@ -81,7 +84,7 @@ import OrbatChart from "@/modules/charteditor/OrbatChart.vue";
 import { injectStrict } from "@/utils";
 import { activeScenarioKey, activeUnitKey } from "@/components/injects";
 import SimpleBreadcrumbs from "@/components/SimpleBreadcrumbs.vue";
-import { BreadcrumbItem } from "@/components/types";
+import { BreadcrumbItem, MenuItemData } from "@/components/types";
 import OrbatChartSettings from "@/modules/charteditor/OrbatChartSettings.vue";
 import {
   OnBranchClickCallback,
@@ -91,6 +94,9 @@ import {
 import { ChartTab, ChartTabs } from "@/modules/charteditor/constants";
 import ToggleField from "@/components/ToggleField.vue";
 import ResizablePanel from "@/components/ResizablePanel.vue";
+import DotsMenu from "@/components/DotsMenu.vue";
+import { promiseTimeout } from "@vueuse/core";
+import FileSaver from "file-saver";
 
 const rootUnitStore = useRootUnitStore();
 const options = useChartSettingsStore();
@@ -123,6 +129,7 @@ const ORBAT_TAB = 0;
 const SETTINGS_TAB = 1;
 
 const selectedTab = ref(ORBAT_TAB);
+const isInteractive = ref(true);
 
 function changeTab(index: number) {
   selectedTab.value = index;
@@ -153,4 +160,71 @@ const onBranchClick: OnBranchClickCallback = (parentId, levelNumber) => {
   changeTab(SETTINGS_TAB);
   currentTab.value = ChartTabs.Branch;
 };
+
+const doSVGDownload = async () => {
+  const origValue = isInteractive.value;
+  isInteractive.value = false;
+  await nextTick();
+  downloadElementAsSVG("chartId");
+  await promiseTimeout(1000);
+  isInteractive.value = origValue;
+};
+
+const doPNGDownload = async () => {
+  const origValue = isInteractive.value;
+  isInteractive.value = false;
+  await nextTick();
+  downloadSvgAsPng("chartId", width.value, height.value);
+  await promiseTimeout(1000);
+  isInteractive.value = origValue;
+};
+
+function downloadSvgAsPng(elementId: string, width: number, height: number) {
+  let svgElement = document.getElementById(elementId);
+  if (!svgElement) return;
+  // need this for Firefox (https://stackoverflow.com/questions/28690643/firefox-error-rendering-an-svg-image-to-html5-canvas-with-drawimage)
+  const savedWidth = svgElement.getAttribute("width") || "";
+  const savedHeight = svgElement.getAttribute("height") || "";
+  const scaleFactor = 2;
+
+  svgElement.setAttribute("width", `${width * scaleFactor}px`);
+  svgElement.setAttribute("height", `${height * scaleFactor}px`);
+  const svgBlob = new Blob([new XMLSerializer().serializeToString(svgElement)], {
+    type: "image/svg+xml",
+  });
+
+  const canvas = document.createElement("canvas");
+  canvas.width = width * scaleFactor;
+  canvas.height = height * scaleFactor;
+  const ctx = canvas.getContext("2d");
+  if (!ctx) return;
+  const objectURL = URL.createObjectURL(svgBlob);
+  const image = new Image();
+
+  image.onload = function () {
+    ctx.drawImage(image, 0, 0);
+    canvas.toBlob((blob) => blob && FileSaver(blob, "orbat-chart.png"));
+    URL.revokeObjectURL(objectURL);
+    svgElement?.setAttribute("width", savedWidth);
+    svgElement?.setAttribute("height", savedHeight);
+  };
+
+  image.src = objectURL;
+}
+
+function downloadElementAsSVG(elementId: string) {
+  let svgElement = document.getElementById(elementId);
+  if (!svgElement) return;
+  FileSaver.saveAs(
+    new Blob([new XMLSerializer().serializeToString(svgElement)], {
+      type: "image/svg+xml",
+    }),
+    "orbat-chart.svg"
+  );
+}
+
+const menuItems: MenuItemData<Function>[] = [
+  { label: "Download as SVG", action: doSVGDownload },
+  { label: "Download as PNG", action: doPNGDownload },
+];
 </script>